// src/types/query.ts
export interface Query {
    date?: number
    jobcard?: number
    name?: number
    item?: number
    quantity?: number
    printing?: number
    corrogation_b2b?: number
    b2b_date?: number
    corrogation_b2b_done?: number
    b2b_date_done?: number
    punching?: number
    pasting?: number
    sent?: number
    stock?: number
}